refactor(sets): rename SetsTab folder to VoteTab
The tab's default label is "Vote" (matching ScheduleTab/InfoTab
naming by label, not route key).
