+++ src/pages/EditionView/tabs/VoteTab/EmptyArtistsState.tsx
