+++ src/pages/EditionView/tabs/VoteTab/MultiArtistSetCard.tsx
