+++ src/pages/EditionView/tabs/VoteTab/SetCard/MultiArtistInfo.tsx
