+++ src/pages/EditionView/tabs/VoteTab/SetCard/SetDescription.tsx
