+++ src/pages/EditionView/tabs/VoteTab/SetCard/SetHeader.tsx
