+++ src/pages/EditionView/tabs/VoteTab/SetCard/SetImage.tsx
