+++ src/pages/EditionView/tabs/VoteTab/SetCard/SetMetadata.tsx
