+++ src/pages/EditionView/tabs/VoteTab/SetCard/SetRatingButtons.tsx
