+++ src/pages/EditionView/tabs/VoteTab/SetCard/SetVotingButtons.tsx
