+++ src/pages/EditionView/tabs/VoteTab/SetCard/SocialPlatformLink.tsx
