+++ src/pages/EditionView/tabs/VoteTab/SetCard/SocialPlatformLinkList.tsx
