+++ src/pages/EditionView/tabs/VoteTab/SetCard/SocialPlatformUtils.tsx
