+++ src/pages/EditionView/tabs/VoteTab/SetCard/social-platform-logos.json
