+++ src/pages/EditionView/tabs/VoteTab/SetListItem.tsx
