+++ src/pages/EditionView/tabs/VoteTab/SetsPanel.tsx
