+++ src/pages/EditionView/tabs/VoteTab/SingleArtistSetCard.tsx
