+++ src/pages/EditionView/tabs/VoteTab/filters/GroupFilterDropdown.tsx
